requirements.txt
annotated-types==0.6.0
anyio==3.7.1
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.5
dnspython==2.4.2
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fastapi==0.104.1
h11==0.14.0
idna==3.4
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==4.6.0
python-deriv-api==0.1.6
python-dotenv==1.0.0
reactivex==4.0.4
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
websockets==10.3
